refactor: Splits temporal-service go files to make it cleaner.
The goal here is gradual improvement of readability to better support
things. An effort was done to not modify interfaces or code.

New files are:

* `process_manager.go`: handles creating processes for goose actions.
* `goose_workflow.go`: creates temporal workflows that trigger goose processes.
* `schedule.go`: Everything related to scheduling of goose workflows.
* `service.go`: Code related to creating a service that manages all the
  above.
* `main.go`: The code that handles integrating with the environment,
  configurations, signals, etc. as well as initiating everything from a command.
